Fixed app config
