+++ README.md
@@ -1,4 +1,27 @@
+# Typescript RestAPI with MongoDB
+Basic RestAPI with Typescript, Mongoose, Authentication
 
+
+## Requirements
+[![NPM Version](https://img.shields.io/badge/npm-10.2.3-blue.svg)](https://www.npmjs.com/)
+
+[![Node.js Version](https://img.shields.io/badge/Node.js-20.10.0-brightgreen.svg)](https://nodejs.org/)
+
+## How to use
+
+### Download or clone this repo
+
+```
+ git clone https://github.com/Kriz1618/typescript-api.git
+ cd typescript-api
+```
+
+### Install
+
+```
+nvm use
+npm install
+```
 
 ## Steps:
 
@@ -11,3 +34,7 @@
 * create the `index.ts` file and configure the server
 * install mongoDB module `npm i mongoose`
 * install mongoDB module types `npm i @types/mongoose`
+
+
+#
+![TypeScript](https://badges.frapsoft.com/typescript/code/typescript.png?v=101)
