+++ package.json
@@ -5,6 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon",
+    "lint": "eslint --ext .ts .",
+    "lint:fix": "eslint --ext .ts --fix .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -19,6 +21,9 @@
     "@types/lodash": "^4.14.202",
     "@types/mongoose": "^5.11.97",
     "@types/morgan": "^1.9.9",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
+    "eslint": "^8.56.0",
     "nodemon": "^3.0.3",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
